Move all tests to the tests dir
