+++ src/decode.rs
@@ -363,59 +363,3 @@ impl fmt::Display for Error {
         }
     }
 }
-// Subset of test cases from https://github.com/cryptocoinjs/base-x/blob/master/test/fixtures.json
-#[cfg(test)]
-mod tests {
-    use decode;
-    use decode::Error;
-
-    #[test]
-    fn tests() {
-        for &(val, s) in super::super::TEST_CASES.iter() {
-            assert_eq!(val.to_vec(), decode(s).into_vec().unwrap());
-        }
-    }
-
-    #[test]
-    fn test_small_buffer_err() {
-        let mut output = [0; 2];
-        assert_eq!(decode("a3gV").into(&mut output), Err(Error::BufferTooSmall));
-    }
-
-    #[test]
-    fn test_invalid_char() {
-        let sample = "123456789abcd!efghij";
-
-        assert_eq!(
-            decode(sample).into_vec().unwrap_err(),
-            Error::InvalidCharacter { character: '!', index: 13 });
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "check")]
-mod test_check{
-    use decode;
-    use decode::Error;
-
-    #[test]
-    fn test_check(){
-        for &(val, s) in super::super::CHECK_TEST_CASES.iter() {
-            assert_eq!(val.to_vec(), decode(s).with_check(None).into_vec().unwrap());
-        }
-
-        for &(val, s) in super::super::CHECK_TEST_CASES[1..].iter() {
-            assert_eq!(val.to_vec(), decode(s).with_check(Some(val[0])).into_vec().unwrap());
-        }
-    }
-
-    #[test]
-    fn test_check_ver_failed() {
-        let d = decode("K5zqBMZZTzUbAZQgrt4")
-            .with_check(Some(0x01))
-            .into_vec();
-
-        assert!(d.is_err());
-        assert_matches!(d.unwrap_err(), Error::InvalidVersion { .. });
-    }
-}
